test: 검증 책임의 이동에 맞게 테스트 이동.
